Final upload for this repository
I am officially retiring this iteration of the project. I lost track of it after a while and the code quality went down hill fast. There's still some good code to be found in here, but a lot of bad code as well. I look forward to reworking this project and hopefully completing it.